+++ Sources/Extensions/SwiftStdlib/IntExtensions.swift
@@ -160,6 +160,11 @@ public extension Int {
 		}
 		return romanValue
 	}
+    
+  /// SwifterSwift: Rounds to the closest multiple of n
+  public func roundToNearest(_ n: Int) -> Int {
+      return n == 0 ? self : Int(round(Double(self) / Double(n))) * n
+  }
 
 }